refactor: 매출 화면 컴포넌트 분리
